refactor(rom-browser): move RomBrowserApp tests into app/tests.rs (#2828)
Relocate the inline 38-test #[cfg(test)] mod tests into a sibling
app/tests.rs (declared from mod.rs). Tests stay in-module; no test
logic or count change.
